src/support-agent/pyproject.toml
[tool.poetry]
name = "support-agent"
version = "0.1.0"
description = ""
authors = ["wojciecham <wojciech.zawadzki@codefit.pl>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
langgraph = "^0.2.60"
langchain = "^0.3.13"
langchain-openai = "^0.2.14"
pydantic = "^2.10.4"
fastapi = "^0.115.6"
uvicorn = "^0.34.0"
pillow = "^11.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
